rendering and viz stuff, prolly broke sonification scripts

+++ core/src/model/types.h
@@ -1,4 +1,4 @@
-typedef float rk_real;
+typedef double rk_real;
 
 /* These parameters affect accuracy (hence speed)*/
 #define N 25 /* Number of host and parasite phenotypes */
